this is my all project and this is my last commit

+++ routes/admin.js
@@ -4,10 +4,7 @@ const adminDb = require('../db-setup/admin')
 const multer = require('multer')
 const path = require('path')
 const fs = require('fs')
-const async = require('hbs/lib/async')
-
-
-// const upload = multer({ dest: 'uploads/' })
+const { Db } = require('mongodb')
 let storage = multer.diskStorage({
   destination: (req, file, cb) => {
     cb(null, './public/uploads')
@@ -21,19 +18,41 @@ let storage = multer.diskStorage({
 })
 const upload = multer({ storage: storage })
 /* GET users listing. */
-router.get('/', function (req, res, next) {
+router.get('/', async function (req, res, next) {
+  if (req.session.admin) {
+    let revenue = await adminDb.revenue()
+    let orders = await adminDb.orders()
+    let user = await adminDb.users()
+
+    let arrRevenue = await adminDb.getRevenue()
+    res.render('admin/dashboard', {
+      admin: true,
+      revenue,
+      orders,
+      user,
+      arrRevenue,
+    })
+  } else {
+    res.redirect('/admin/login')
+  }
+})
+router.get('/login', (req, res) => {
   res.render('admin/login', { admin: true })
 })
 router.post('/login', (req, res) => {
-  if (
-    req.body.email === 'captian.8089@gmail.com' &&
-    req.body.password === '1212'
-  ) {
-    res.redirect('/admin/all-products')
-  }
+  adminDb.doLogin(req.body).then((response) => {
+    if (response.status) {
+      req.session.admin = true
+      res.redirect('/admin/dashboard')
+    }
+  })
 })
 router.get('/add-product', (req, res) => {
-  res.render('admin/add-product', { admin: true })
+  if (req.session.admin) {
+    res.render('admin/add-product', { admin: true })
+  } else {
+    res.redirect('/admin/')
+  }
 })
 router.post('/add-product', upload.array('images', 3), async (req, res) => {
   let file = req.files
@@ -56,21 +75,33 @@ router.post('/add-product', upload.array('images', 3), async (req, res) => {
   res.redirect('/admin/add-product')
 })
 router.get('/all-products', (req, res) => {
-  adminDb.getAllProduct().then((products) => {
-    res.render('admin/all-products', { products, admin: true })
-  })
+  if (req.session.admin) {
+    adminDb.getAllProduct().then((products) => {
+      res.render('admin/all-products', { products, admin: true })
+    })
+  } else {
+    res.redirect('/admin/')
+  }
 })
 router.get('/delete-product/:id', (req, res) => {
-  let proId = req.params.id
-  adminDb.deleteProduct(proId).then((response) => {
-    res.redirect('/admin/all-products')
-  })
+  if (req.session.admin) {
+    let proId = req.params.id
+    adminDb.deleteProduct(proId).then((response) => {
+      res.redirect('/admin/all-products')
+    })
+  } else {
+    res.redirect('/admin/')
+  }
 })
 router.get('/edit-product/:id', async (req, res) => {
-  let proId = req.params.id
-  let product = await adminDb.getProductDetails(proId)
+  if (req.session.admin) {
+    let proId = req.params.id
+    let product = await adminDb.getProductDetails(proId)
 
-  res.render('admin/edit-product', { product, admin: true })
+    res.render('admin/edit-product', { product, admin: true })
+  } else {
+    res.redirect('/admin/')
+  }
 })
 router.post(
   '/edit-product/:id',
@@ -97,33 +128,94 @@ router.post(
   },
 )
 router.get('/all-users', async (req, res) => {
-  let users = await adminDb.getUsers()
+  if (req.session.admin) {
+    let users = await adminDb.getUsers()
 
-  res.render('admin/all-users', { users, admin: true })
+    res.render('admin/all-users', { users, admin: true })
+  } else {
+    res.redirect('/admin/')
+  }
 })
 router.get('/block-user/:id', (req, res) => {
-  adminDb.blockUser(req.params.id).then(() => {
-    req.session.blocked = true
-    res.redirect('/admin/all-users')
-  })
+  if (req.session.admin) {
+    adminDb.blockUser(req.params.id).then(() => {
+      req.session.blocked = true
+      res.redirect('/admin/all-users')
+    })
+  } else {
+    res.redirect('/admin/')
+  }
 })
 router.get('/unblock-user/:id', (req, res) => {
-  adminDb.unBlockUser(req.params.id).then(() => {
-    res.redirect('/admin/all-users')
+  if (req.session.admin) {
+    adminDb.unBlockUser(req.params.id).then(() => {
+      res.redirect('/admin/all-users')
+    })
+  } else {
+    res.redirect('/admin/')
+  }
+})
+router.get('/user-orders/:id', async (req, res) => {
+  if (req.session.admin) {
+    let order = await adminDb.userOrders(req.params.id)
+    let total = await adminDb.totalByTheUser(req.params.id)
+
+    res.render('admin/user-orders', { order, total, admin: true })
+  } else {
+    res.redirect('/admin/')
+  }
+})
+router.get('/all-orders', (req, res) => {
+  if (req.session.admin) {
+    adminDb.allOrders().then((allOrders) => {
+      console.log(allOrders)
+      res.render('admin/all-orders', { allOrders, admin: true })
+    })
+  } else {
+    res.redirect('/admin/')
+  }
+})
+router.get('/coupen', async (req, res) => {
+  if (req.session.admin) {
+    let coupen = await adminDb.getCoupen()
+
+    res.render('admin/coupen', { coupen, admin: true })
+  } else {
+    res.redirect('/admin/')
+  }
+})
+router.post('/coupen', (req, res) => {
+  console.log(req.body)
+  adminDb.coupen(req.body).then((response) => {
+    res.redirect('/admin/coupen')
   })
 })
-router.get('/user-orders/:id',async(req, res) => {
- let order =await adminDb.userOrders(req.params.id)
- let total = await adminDb.totalByTheUser(req.params.id)
- console.log(order);
-  res.render('admin/user-orders',{order,total})
-})
-router.get('/all-orders',(req, res) => {
-  adminDb.allOrders().then((allOrders) => {
-    console.log(allOrders);
-    res.render('admin/all-orders',{allOrders, admin:true})
+router.post('/coupen-delete', (req, res) => {
+  adminDb.coupenDelete(req.body).then((response) => {
+    res.json(response)
   })
+})
+router.get('/dashboard', async (req, res) => {
+  if (req.session.admin) {
+    let revenue = await adminDb.revenue()
+    let orders = await adminDb.orders()
+    let user = await adminDb.users()
+
+    let arrRevenue = await adminDb.getRevenue()
+    res.render('admin/dashboard', {
+      admin: true,
+      revenue,
+      orders,
+      user,
+      arrRevenue,
+    })
+  } else {
+    res.redirect('/admin/')
+  }
+})
+router.get('/logout', (req, res) => {
+  req.session.admin = false
+  res.redirect('/admin/')
 })
 
 module.exports = router
-
